Android: Make sure backend API is called on right thread
The Java code lives in the Android UI thread, so we need to use
invokeMethod() to make sure we call into the right thread. This
simplifies the code because we don't have to worry about
synchronization, and it also makes it possible to create children
of the backend and use regular QObject memory management.

Change-Id: I5d1673ffc9cd6f0c9e0456eade39ef2d4dfd6701
Reviewed-by: Eskil Abrahamsen-Blomfeldt <eskil.abrahamsen-blomfeldt@digia.com>