Make QInAppStore accessible from backend
The backend sometimes needs to call back to the QInAppStore. This
is specifically needed in the Android backend when restoring
purchases to access the registered products. The alternative
is to move the list of products into the backend which would make
the call graph of product registration even more complex, and
having this pointer might be useful later on as well.

Change-Id: I48e495fb947b467acaf64ca35b651da7c178da20
Reviewed-by: Andy Nichols <andy.nichols@digia.com>