Android: Fix restoring purchases
First of all, in order to be able to emit a transaction for existing
purchases, we need to retain the purchase info from when we queried
the purchase. We also need to add the same data to that map when
the item is purchased, otherwise we would not be able to restore
purchases that was done during the same execution.

In addition, we need to register the transaction as finalized
for restore transactions. To make this work, we overwrite the
list of finalized products each time. This also makes sure the
file is always the optimal size.

Change-Id: Ic2f7cae7f8aa4cbea0e8805ffca1c1d862686870
Reviewed-by: Eskil Abrahamsen-Blomfeldt <eskil.abrahamsen-blomfeldt@digia.com>