Make it possible to create products outside store
The current API did not allow for the use case of creating
products from a model in a convenient way, since there was
no way of setting the store pointer except to set the
parent of the objects, which is hacky.

This adds explicitly setting the store as an option.

Change-Id: Id98a9d646afa8818122b73e8bf4aac61e90a4cab
Reviewed-by: Andy Nichols <andy.nichols@digia.com>