Remove ConsumableProduct and UnlockableProduct subtypes
When instantiating products dynamically, e.g from a model, this
design breaks down. Since it doesn't have any purpose except
for readability and actually makes the documentation a bit
more awkward to read, I've just removed the ConsumableProduct
and UnlockableProduct and replaced them with an enum in the
Product type instead.

Change-Id: I448d063fafe1aba39daf98a7a8ad4e168564a565
Reviewed-by: Andy Nichols <andy.nichols@digia.com>