Android: Fix finalization check for consumables
If a consumable item is registered as purchased in the store, it
means it has not been finalized. The condition was reversed here,
causing consumables to never be regarded as unfinalized. This
meant that if a consumable was purchased and the transaction for
some reason was not finalized, the transaction would not be
delivered again the next time it was registered, as it should have
been.

Change-Id: Ife2cb04cbb4c994673cac7c25680881672b859aa
Reviewed-by: Andy Nichols <andy.nichols@digia.com>