Make MODULE_VERSION=1.0.0
This will only affect the versions of the library files, but it
makes sense to detach it from Qt versioning and use the actual
version from the Qt Quick module instead, both because this is
easier to track and because we won't have to update it with
every release.

This is also how other enterprise add-ons do it.

Change-Id: I4f2f4a99dc27bc6fe5f46aebcbca4cfce1aa8e68
Reviewed-by: Johanna Äijälä <johanna.aijala@theqtcompany.com>
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>