Android: Fix false negatives in apps with many products
In an application with many products (e.g. 50), some would
randomly fail to be requested, sometimes with remote exceptions
and other times with regular failures. It could be that the
service object is not thread-safe, but it seems to be easier
to reproduce on low-end devices, so a plausible explanation
is that the amount of simultaneous requests drains all
resources. Synchronizing access will slow it down a little
bit, but this can be fixed in a follow-up patch where we
request all products in a single request instead of
individually.

Change-Id: I19ca5aa43dd780fdd264f179f8a8ece603215a3e
Task-number: QTRD-3506
Reviewed-by: Christian Stromme <christian.stromme@theqtcompany.com>