Android: Batch requests for products
To speed up product requests, we can batch them in bundles of
at most 20 SKUs (according to the documentation).

To do this, we add a new virtual function to the back-end which
allows sending in the whole list of pending products. The
default implementation of this just calls the regular
queryProduct() function so that behavior stays the same
on iOS.

Change-Id: I093e254792d23ed39199392748b7163925856081
Task-number: QTRD-3506
Reviewed-by: Christian Stromme <christian.stromme@theqtcompany.com>