Add Product.resetStatus() method
When the product query for some reason fails, the only way to
re-request the product currently is to use instantiators, delete
the object and then recreate it.

Instead, we add explicit API to accomplish this. Since we
want to be able to do this from the onStatusChanged handler,
it required a small change to the order of events in Android,
the item needs to be removed from the pending list before the
status can be reset, so the signal that the query failed has
to be the last thing that happens.

Task-number: QTRD-3506
Change-Id: I072ff9c965986b00b2b7557535e19af1b47b1b51
Reviewed-by: Venugopal Shivashankar <venugopal.shivashankar@theqtcompany.com>
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>