Revert "Bump version to 1.1"
This reverts commit 51b93c9afa2c8b5797403bd686af6b2b25544dfb and
sets the module version to 1.0.1 instead. We don't need to update
the minor version for adding methods, and furthermore, there does
not seem to be any way to increase the minor version of a QML
import without adding at least one type, since you have to
register at type with the new version number for it to be
available, and you can't do this with existing types, since
they will then become unavailable in the previous version.

Change-Id: I3933b2c17389b8bc130ece566196954777bec566
Reviewed-by: Simon Hausmann <simon.hausmann@theqtcompany.com>