Make module version 5.6.0
In 933318699b19d469ea85ab0d90f77efdb9940975 we detached the
module version from Qt because this was consistent with how
other commercial add-ons were done, but now we intend to add
the module to qt5.git, so we need to go back to following
Qt versions. As before, this only affects the library
files and should not cause any trouble with upgrading.

Change-Id: I0c9a4e133b05eeb07fabdb5916bf5c4cd674dd96
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>