Fix crash when requerying product before previous completed
If a product was queried twice in a row, so quickly that the first
one was still pending when the second query was made, then we
would have unexpected null pointers in the array of identifiers
that we pass into Java. This would cause a crash. Instead, we
remove all identifiers which are already pending before creating
the array so that it will not have any holes.

Task-number: QTBUG-52330
Change-Id: I643835ea224dac969d29ad6572f224a068724d41
Reviewed-by: Andy Nichols <andy.nichols@qt.io>