Android: Fix test purchases after changes in Google Play
After June 20th, Google Play no longer generates order IDs for
test purchases, causing an exception to occur in Qt Purchasing
when making such purchases and preventing the transaction from
being communicated to the application. The orderId and timestamp
are not technically required information, so in the case where
they are not available, we just ignore it and use defaults
instead.

[ChangeLog][Android] Fixed test purchases after changes in
Google Play that removed the order ID for such transactions.

Task-number: QTBUG-54518
Change-Id: If8cd8e4dc76329e572556988b3e65485920c4e0d
Reviewed-by: BogDan Vatra <bogdan@kdab.com>