Android: More informative error string in some cases
Some devices will fail synchronously when trying to create
the buy intent, instead of calling handleActivityResult()
with the result.

When trying to purchase an item already owned by the user
on a such device, you would get a purchase failed signal
with a very generic error string, as we assumed this could
only be caused by a system error. This patch improves the
error message by including a description of the response code
in the cases where the system returns something usable here.

[ChangeLog][Android] Added more informative error messages when
the purchasing subsystem was unable to create the buy intent.

Task-number: QTBUG-57700
Change-Id: I6328b38b33edf4c1c53d2465bf2205c970bf7e78
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>