Update commercial license headers
Updated header.COMM to the files in tqtc-qtpurchasing.
Examples, tests and documentation files are not
updated.

The commercial license header may contain some
additional lines so that its line count equals
with the earlier license header. Reason for this
is that some autotests use hard coded line numbers
and a change in the line count causes failures in
tests.

Task-number: QTQAINFRA-4150
Change-Id: I29e2cf34dd936206efb48a95c07cc37c729ed63d
Reviewed-by: Antti Kokko <antti.kokko@qt.io>